RF: move api and refactor Game.js;
